Consolidate Select logic in useSelect hook
